[CSA-26] Added common package for json serialization and deserialization between angle and voltage monitoring
