+++ monitoring/angle-monitoring/src/main/java/com/farao_community/farao/monitoring/angle_monitoring/json/AngleMonitoringResultSerializer.java
@@ -8,18 +8,19 @@
 package com.farao_community.farao.monitoring.angle_monitoring.json;
 
 import com.farao_community.farao.data.crac_api.Contingency;
-import com.farao_community.farao.data.crac_api.State;
-import com.farao_community.farao.data.crac_api.network_action.NetworkAction;
 import com.farao_community.farao.monitoring.angle_monitoring.AngleMonitoringResult;
+import com.farao_community.farao.monitoring.monitoring_common.json.MonitoringCommonSerializer;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 
 import java.io.IOException;
-import java.util.*;
+import java.util.Comparator;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import static com.farao_community.farao.monitoring.angle_monitoring.json.JsonAngleMonitoringResultConstants.*;
+import static com.farao_community.farao.monitoring.monitoring_common.json.JsonCommonMonitoringResultConstants.*;
 
 /**
  * @author Godelaine de Montmorillon {@literal <godelaine.demontmorillon at rte-france.com>}
@@ -42,7 +43,7 @@ public class AngleMonitoringResultSerializer extends JsonSerializer<AngleMonitor
         jsonGenerator.writeEndArray();
         // APPLIED_CRAS
         jsonGenerator.writeArrayFieldStart(APPLIED_CRAS);
-        serializeAppliedCras(angleMonitoringResult, jsonGenerator);
+        MonitoringCommonSerializer.serializeAppliedRas(angleMonitoringResult.getAppliedCras(), jsonGenerator);
         jsonGenerator.writeEndArray();
 
         jsonGenerator.writeEndObject();
@@ -58,24 +59,6 @@ public class AngleMonitoringResultSerializer extends JsonSerializer<AngleMonitor
             }
             jsonGenerator.writeStringField(CNEC_ID, angleResult.getAngleCnec().getId());
             jsonGenerator.writeNumberField(QUANTITY, Math.round(10.0 * angleResult.getAngle()) / 10.0);
-            jsonGenerator.writeEndObject();
-        }
-    }
-
-    private void serializeAppliedCras(AngleMonitoringResult angleMonitoringResult, JsonGenerator jsonGenerator) throws IOException {
-        for (Map.Entry<State, Set<NetworkAction>> entry : angleMonitoringResult.getAppliedCras().entrySet().stream().sorted(Comparator.comparing(e -> e.getKey().getId()))
-                .collect(Collectors.toList())) {
-            jsonGenerator.writeStartObject();
-            jsonGenerator.writeStringField(INSTANT, entry.getKey().getInstant().toString());
-            Optional<Contingency> optContingency = entry.getKey().getContingency();
-            if (optContingency.isPresent()) {
-                jsonGenerator.writeStringField(CONTINGENCY, optContingency.get().getId());
-            }
-            jsonGenerator.writeArrayFieldStart(REMEDIAL_ACTIONS);
-            for (NetworkAction networkAction : entry.getValue().stream().sorted(Comparator.comparing(NetworkAction::getId)).collect(Collectors.toList())) {
-                jsonGenerator.writeString(networkAction.getId());
-            }
-            jsonGenerator.writeEndArray();
             jsonGenerator.writeEndObject();
         }
     }
