+++ monitoring/monitoring-common/pom.xml
@@ -9,18 +9,13 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>farao-voltage-monitoring</artifactId>
+    <artifactId>farao-monitoring-common</artifactId>
     <packaging>jar</packaging>
-    <name>Voltage monitoring</name>
-    <description>Voltage monitoring features</description>
+    <name>Monitoring common</name>
+    <description>Common voltage and angle monitoring features</description>
 
     <dependencies>
         <!-- Compile dependencies -->
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>farao-monitoring-common</artifactId>
-            <version>${project.version}</version>
-        </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>farao-crac-api</artifactId>
