+++ monitoring/pom.xml
@@ -18,5 +18,6 @@
     <modules>
         <module>angle-monitoring</module>
         <module>voltage-monitoring</module>
+        <module>monitoring-common</module>
     </modules>
 </project>
