+++ monitoring/voltage-monitoring/src/main/java/com/farao_community/farao/monitoring/voltage_monitoring/json/VoltageMonitoringResultDeserializer.java
@@ -9,22 +9,23 @@ package com.farao_community.farao.monitoring.voltage_monitoring.json;
 
 import com.farao_community.farao.commons.FaraoException;
 import com.farao_community.farao.data.crac_api.Crac;
-import com.farao_community.farao.data.crac_api.Instant;
 import com.farao_community.farao.data.crac_api.State;
 import com.farao_community.farao.data.crac_api.cnec.VoltageCnec;
 import com.farao_community.farao.data.crac_api.network_action.NetworkAction;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonToken;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
+import com.farao_community.farao.monitoring.monitoring_common.json.MonitoringCommonDeserializer;
 import com.farao_community.farao.monitoring.voltage_monitoring.ExtremeVoltageValues;
 import com.farao_community.farao.monitoring.voltage_monitoring.VoltageMonitoringResult;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.JsonToken;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
 
 import java.io.IOException;
-import java.util.*;
-import java.util.stream.Collectors;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Set;
 
+import static com.farao_community.farao.monitoring.monitoring_common.json.JsonCommonMonitoringResultConstants.*;
 import static com.farao_community.farao.monitoring.voltage_monitoring.json.JsonVoltageMonitoringResultConstants.*;
 
 /**
@@ -65,7 +66,7 @@ public class VoltageMonitoringResultDeserializer extends JsonDeserializer<Voltag
                 readVoltageValues(jsonParser, extremeVoltageValues);
             } else if (jsonParser.getCurrentName().equals(APPLIED_RAS)) {
                 jsonParser.nextToken();
-                readAppliedRas(jsonParser, appliedRas);
+                MonitoringCommonDeserializer.readAppliedRas(jsonParser, appliedRas, crac);
             } else {
                 throw new FaraoException(String.format("Unexpected field %s in %s", jsonParser.getCurrentName(), VOLTAGE_MONITORING_RESULT));
             }
@@ -116,59 +117,5 @@ public class VoltageMonitoringResultDeserializer extends JsonDeserializer<Voltag
             }
             voltageValues.put(voltageCnec, new ExtremeVoltageValues(Set.of(min, max)));
         }
-    }
-
-    private void readAppliedRas(JsonParser jsonParser, Map<State, Set<NetworkAction>> appliedRas) throws IOException {
-        while (jsonParser.nextToken() != JsonToken.END_ARRAY) {
-            String contingencyId = null;
-            Instant instant = null;
-            Set<String> remedialActionIds = new HashSet<>();
-            while (!jsonParser.nextToken().isStructEnd()) {
-                switch (jsonParser.currentName()) {
-                    case INSTANT:
-                        instant = deserializeInstant(jsonParser.nextTextValue());
-                        break;
-                    case CONTINGENCY:
-                        contingencyId = jsonParser.nextTextValue();
-                        break;
-                    case REMEDIAL_ACTIONS:
-                        jsonParser.nextToken();
-                        remedialActionIds = jsonParser.readValueAs(new TypeReference<HashSet<String>>() {
-                        });
-                        break;
-                    default:
-                        throw new FaraoException(String.format(UNEXPECTED_FIELD_ERROR, jsonParser.currentName(), REMEDIAL_ACTIONS));
-                }
-            }
-            if (instant == null) {
-                throw new FaraoException(String.format("Instant must be defined in %s", REMEDIAL_ACTIONS));
-            }
-            // Get network Actions from string
-            State state = getState(instant, contingencyId);
-            if (appliedRas.containsKey(state)) {
-                throw new FaraoException(String.format("State with instant %s and contingency %s has previously been defined in %s", instant.toString(), contingencyId, REMEDIAL_ACTIONS));
-            } else {
-                appliedRas.put(state, getNetworkActions(remedialActionIds));
-            }
-        }
-    }
-
-    private State getState(Instant instant, String contingencyId) {
-        if (Objects.isNull(contingencyId)) {
-            if (instant.equals(Instant.PREVENTIVE)) {
-                return crac.getPreventiveState();
-            } else {
-                throw new FaraoException(String.format("No contingency defined with instant %s", instant.toString()));
-            }
-        }
-        State state = crac.getState(contingencyId, instant);
-        if (Objects.isNull(state)) {
-            throw new FaraoException(String.format("State with instant %s and contingency %s does not exist in CRAC", instant.toString(), contingencyId));
-        }
-        return state;
-    }
-
-    private Set<NetworkAction> getNetworkActions(Set<String> ids) {
-        return crac.getNetworkActions().stream().filter(networkAction -> ids.contains(networkAction.getId())).collect(Collectors.toSet());
     }
 }
