+++ monitoring/voltage-monitoring/src/main/java/com/farao_community/farao/monitoring/voltage_monitoring/json/VoltageMonitoringResultSerializer.java
@@ -7,10 +7,8 @@
 
 package com.farao_community.farao.monitoring.voltage_monitoring.json;
 
-import com.farao_community.farao.data.crac_api.Contingency;
-import com.farao_community.farao.data.crac_api.State;
 import com.farao_community.farao.data.crac_api.cnec.VoltageCnec;
-import com.farao_community.farao.data.crac_api.network_action.NetworkAction;
+import com.farao_community.farao.monitoring.monitoring_common.json.MonitoringCommonSerializer;
 import com.farao_community.farao.monitoring.voltage_monitoring.ExtremeVoltageValues;
 import com.farao_community.farao.monitoring.voltage_monitoring.VoltageMonitoringResult;
 import com.fasterxml.jackson.core.JsonGenerator;
@@ -20,10 +18,9 @@ import com.fasterxml.jackson.databind.SerializerProvider;
 import java.io.IOException;
 import java.util.Comparator;
 import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
 import java.util.stream.Collectors;
 
+import static com.farao_community.farao.monitoring.monitoring_common.json.JsonCommonMonitoringResultConstants.*;
 import static com.farao_community.farao.monitoring.voltage_monitoring.json.JsonVoltageMonitoringResultConstants.*;
 
 /**
@@ -32,7 +29,6 @@ import static com.farao_community.farao.monitoring.voltage_monitoring.json.JsonV
 public class VoltageMonitoringResultSerializer extends JsonSerializer<VoltageMonitoringResult> {
 
     VoltageMonitoringResultSerializer() {
-
     }
 
     @Override
@@ -47,7 +43,7 @@ public class VoltageMonitoringResultSerializer extends JsonSerializer<VoltageMon
         jsonGenerator.writeEndArray();
         // APPLIED_RAS
         jsonGenerator.writeArrayFieldStart(APPLIED_RAS);
-        serializeAppliedRas(voltageMonitoringResult, jsonGenerator);
+        MonitoringCommonSerializer.serializeAppliedRas(voltageMonitoringResult.getAppliedRas(), jsonGenerator);
         jsonGenerator.writeEndArray();
 
         jsonGenerator.writeEndObject();
@@ -65,23 +61,4 @@ public class VoltageMonitoringResultSerializer extends JsonSerializer<VoltageMon
             jsonGenerator.writeEndObject();
         }
     }
-
-    private void serializeAppliedRas(VoltageMonitoringResult voltageMonitoringResult, JsonGenerator jsonGenerator) throws IOException {
-        for (Map.Entry<State, Set<NetworkAction>> entry : voltageMonitoringResult.getAppliedRas().entrySet().stream().sorted(Comparator.comparing(e -> e.getKey().getId()))
-                .collect(Collectors.toList())) {
-            jsonGenerator.writeStartObject();
-            jsonGenerator.writeStringField(INSTANT, entry.getKey().getInstant().toString());
-            Optional<Contingency> optContingency = entry.getKey().getContingency();
-            if (optContingency.isPresent()) {
-                jsonGenerator.writeStringField(CONTINGENCY, optContingency.get().getId());
-            }
-            jsonGenerator.writeArrayFieldStart(REMEDIAL_ACTIONS);
-            for (NetworkAction networkAction : entry.getValue().stream().sorted(Comparator.comparing(NetworkAction::getId)).collect(Collectors.toList())) {
-                jsonGenerator.writeString(networkAction.getId());
-            }
-            jsonGenerator.writeEndArray();
-            jsonGenerator.writeEndObject();
-        }
-    }
-
 }
